Move to Reference Server & start implementing queue restore
