+++ Applications/ConsoleReferenceClient/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "ConsoleReferenceClient": {
+      "commandName": "Project",
+      "commandLineArgs": "-ds -un sysadmin -up demo"
+    }
+  }
+}
